Update windows top script limit from 30 to 10000
In Windows OS currently, yc agent is transmitting only 30 processes
in the top file to the yc server. Need to enhance the Top script to
transmit all the processes which are running on the Windows server.

Since I don't have a windows machine to test, I am avoiding making a
big change to the script. So, changing 30 to 10000 should give enough
limit to contain all processes.